+++ com.splunk.modularinput.java.ui/src/com/splunk/modularinput/java/ui/WidgetCreation.java
@@ -1,0 +1,8 @@
+package com.splunk.modularinput.java.ui;
+
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
+
+public interface WidgetCreation<T extends Control> {
+	public T init(Composite parent);
+}