+++ CHANGELOG.md
@@ -1,4 +1,27 @@
-# Splunk plug-in for Eclipse
+# Splunk Plug-in for Eclipse
+
+## Version 1.0.1
+
+* Updated Splunk SDK for Java in plugin for version 1.3.
+
+## Version 1.0
+
+* Switched to Maven for building the Splunk Plug-in for Eclipse.
+* Updated logback and log4j libraries (now using log4j 2).
+* Added updated Splunk Logging for Java library.
+* Splunk entries in the new project wizard now have an icon.
+* Removed monitored Java launcher. If you need this functionality, contact
+  devinfo@splunk.com to get access to an altered build that retains the
+  unsupported functionality from the beta release.
+
+## Version 0.8
+
+* Added support for creating modular inputs in Java.
+
+## Version 0.2.1
+
+* Fixed errors in dependency versions so plug-in works on versions of Eclipse back to 3.6.2.
+* Upgraded the Splunk SDK for Java version 1.2.1.
 
 ## Version 0.2